Correction to field

+++ core_backend/models.py
@@ -296,7 +296,7 @@ class Service(ExtendableModel):
     categories = models.ManyToManyField(Category, related_name='services')
     provider = models.ForeignKey(Provider, on_delete=models.CASCADE, related_name='services')
     bill_amount = models.DecimalField(_('billing amount'), max_digits=32, decimal_places=2)
-    bill_rate = models.DecimalField(_('billing rate in seconds'), max_digits=32, decimal_places=2)
+    bill_rate = models.IntegerField(_('billing rate in seconds'))
     bill_type = models.CharField(max_length=3, choices=BillType.choices, default=BillType.PER_MINUTE)
 
     class Meta:
